♻️refactor: API 관련 패키지 생성 및 상수 관리를 위한 ResponseMessage 구현
