+++ backend/src/main/java/codesquard/app/api/errors/exception/CommentMaxLengthExceededException.java
@@ -1,4 +1,4 @@
-package codesquard.app.errors.exception;
+package codesquard.app.api.errors.exception;
 
 public class CommentMaxLengthExceededException extends RuntimeException {
