+++ backend/src/main/java/codesquard/app/api/errors/handler/CommentExceptionHandler.java
@@ -1,4 +1,4 @@
-package codesquard.app.errors.handler;
+package codesquard.app.api.errors.handler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import codesquard.app.errors.exception.CommentMaxLengthExceededException;
-import codesquard.app.errors.response.ApiResponse;
+import codesquard.app.api.errors.exception.CommentMaxLengthExceededException;
+import codesquard.app.api.response.ApiResponse;
 
 @RestControllerAdvice
 public class CommentExceptionHandler {
