+++ backend/src/main/java/codesquard/app/api/errors/handler/ImageExceptionHandler.java
@@ -1,4 +1,4 @@
-package codesquard.app.errors.handler;
+package codesquard.app.api.errors.handler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.multipart.MaxUploadSizeExceededException;
 
-import codesquard.app.errors.exception.EmptyFileException;
-import codesquard.app.errors.response.ApiResponse;
+import codesquard.app.api.errors.exception.EmptyFileException;
+import codesquard.app.api.response.ApiResponse;
 
 @RestControllerAdvice
 public class ImageExceptionHandler {
