+++ backend/src/main/java/codesquard/app/api/response/ResponseMessage.java
@@ -1,0 +1,7 @@
+package codesquard.app.api.response;
+
+public abstract class ResponseMessage {
+
+	public static final String COMMENT_SAVE_SUCCESS = "댓글 등록에 성공했습니다.";
+
+}
